catalog/viewsets.py
from catalog.serializers import BookSerializer
from catalog.models import Book
from rest_framework import viewsets

class BookViewSet(viewsets.ModelViewSet):
    queryset = Book.objects.all()
    serializer_class = BookSerializer