requirements.txt
django
django-bootstrap4
djangorestframework